+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/utils/ReflectUtils.java
@@ -15,6 +15,9 @@
  */
 package io.fabric8.kubernetes.client.utils;
 
+import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.fabric8.kubernetes.api.model.KubernetesResourceList;
+import io.fabric8.kubernetes.api.model.ListMeta;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 
 import java.lang.reflect.InvocationTargetException;
@@ -24,8 +27,11 @@ import java.util.List;
 public class ReflectUtils {
 
   public static ObjectMeta objectMetadata(Object obj) throws ReflectiveOperationException {
+    if (obj instanceof HasMetadata) {
+      return ((HasMetadata) obj).getMetadata();
+    }
     try {
-      Method mdField = obj.getClass().getDeclaredMethod("getMetadata");
+      Method mdField = obj.getClass().getMethod("getMetadata");
       return (ObjectMeta) mdField.invoke(obj);
     } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
       throw new ReflectiveOperationException(e);
@@ -36,16 +42,22 @@ public class ReflectUtils {
     return objectMetadata(obj).getNamespace();
   }
 
-  public static ObjectMeta listMetadata(Object listObj) throws ReflectiveOperationException {
+  public static ListMeta listMetadata(Object listObj) throws ReflectiveOperationException {
+    if (listObj instanceof KubernetesResourceList) {
+      return ((KubernetesResourceList<?>) listObj).getMetadata();
+    }
     try {
-      Method mdField = listObj.getClass().getDeclaredMethod("getMetadata");
-      return (ObjectMeta) mdField.invoke(listObj);
+      Method mdField = listObj.getClass().getMethod("getMetadata");
+      return (ListMeta) mdField.invoke(listObj);
     } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
       throw new ReflectiveOperationException(e);
     }
   }
 
   public static <ApiType> List<ApiType> getItems(Object listObj) throws ReflectiveOperationException {
+    if (listObj instanceof KubernetesResourceList) {
+      return ((KubernetesResourceList) listObj).getItems();
+    }
     try {
       Method getItemsMethod = listObj.getClass().getMethod("getItems");
       return (List<ApiType>) getItemsMethod.invoke(listObj);